Merge branch 'dev' of github.com:msimms/StraenWeb

+++ ActivityHasher.py
@@ -64,10 +64,10 @@ class ActivityHasher(object):
                 else:
                     for datum in self.activity[sensor_type]:
                         if sys.version_info[0] < 3:
-                            time = str(int(datum.keys()[0])).encode('utf-8')
+                            time = str(float(datum.keys()[0])).encode('utf-8')
                             value = self.floatToStr(float(datum.values()[0]))
                         else:
-                            time = str(int(list(datum.keys())[0])).encode('utf-8')
+                            time = str(float(list(datum.keys())[0])).encode('utf-8')
                             value = self.floatToStr(float(list(datum.values())[0]))
 
                         h.update(time)
